pyproject.toml
    [project]
    name = "gerenciamento-python"
    version = "0.1.0"
    description = "Sistema_De_gerenciamento"
    authors = [
        {name = "AndersonHyoukaJoao"}
    ]
    license = {text = "MIT"}
    readme = "README.md"
    requires-python = ">=3.11,<4.0"
    dependencies = [
        "mysql-connector-python (>=9.3.0,<10.0.0)",
        "pandas (>=2.2.3,<3.0.0)",
        "python-dotenv (>=1.1.0,<2.0.0)"
    ]


    [build-system]
    requires = ["poetry-core>=2.0.0,<3.0.0"]
    build-backend = "poetry.core.masonry.api"
